Fix Layout for dynamic pages
Current the layout component relied on the `pageContext` prop to render each page with information provided in the frontmatter. But dynamic pages (created as `.js` files) don't set the `pageContext` prop, causing the layout to fail.

To mitigate this, we provide a default value to the `pageContext` argument and extract values conditionally based on whether the prop was passed or not.