Adjust header display
Given that the site's navbar is fixed positioned, it would overlap content below it as it is removed from normal flow. To mitigate this, wrap the navbar in a box with the same height as the navbar.

Remove the top margin on the page hero wrapper as it is no longer needed. Also adjust the height of wrapper around the navbar at break point to avoid the gap between it and the page hero.