Restrict access to site
This is only a dev requirement so that the site can be tested in a live environment without it being accessible to the public. As it is only a temporary measure, the implementation is not optimal.

There is no easy fix to restrict access to statically generated content in gatsby as it is already rendered before any auth logic can be applied. However, using the `wrapRootElement` api one is able to perform some logic before the page is rendered. Ideally you would want to redirect upon successful login, but it turns out a clean way to do this is not straightforward.