Configure AWS authentication
Authentication configured only to enable login. At this stage, this is deemed sufficient for the purposes of restricting access to the site. Once logged in, navigation of the site is possible but mechanism for logging out is provided at the moment as it is not at the moment clear what value such functionality would add.