Add search functionality
Search begins with the user supplying the query string and upon submit the search page is displayed with the search results from the hydrated site index. The current query string and search results are persisted in state managed by the `SearchContext` which is provided via wrapping of the root element.

At this point filtering of the results is not yet supported and will be added in the near future. Also not supported at the moment is pagination of results. As this is not a huge site, it remains to be seen whether such functionality will really be needed in the end. Still more styling work is required to make the page look and behave as per design.

There is some code duplication which should be refactored for maintainability. This was due to the fact that the search component in the navbar does not lend itself easily to reusability.