Clean up CSS and code refactoring
The CSS has been clean up and reorganized to clarify the purpose of each style rule. The `styles` directory now contains all the styles that are not encapsulated inside components. Styles are grouped thematically: layout, text, images, lists, tables, etc. There are enough comments to further clarify each rule and to aid in future modifications if and when required.

Emphasis was put on using semantic selectors as much as possible. Class selectors are only used in very few cases within the layout styles. This also enabled to clean up the content of the pages to remain as close to pure markdown as possible.

Instead of using two separate components to render custom titles, the `Title` component has been refactored to cater for two use cases determined by the passed in props.

The `data/backup` folder contains old version of changed files for reference.