Suppress eslint warnings
React `useEffect()` hook usage dictates explicitly declaring values
upon which the effect depends in a dependency array passed as the
second argument. In this particular case, eslint is complaining about
a function used to update state, but unless we figure out how to do
that elsewhere, there does not seem to be any easy way out. Adding the
function to the dependency array would lead to an infinite loop unless
it is memoized with useCallback() hook, which seems overkill for this
particular use case. So, for the time being, we opt to suppress the
warning instead, as there don't seem to be any adverse side effect in
doing so.